machbase/neo-web#536 - update overlap code
machbase/neo-web#536 - update overlap code
