+++ src/components/tagAnalyzer/OverlapModal.scss
@@ -1,4 +1,4 @@
-.compare-cover {
+.Overlap-cover {
     width: 100vw;
     height: 100vh;
     z-index: 998;
@@ -7,7 +7,7 @@
     top: 0;
     left: 0;
 }
-.compare-modal {
+.Overlap-modal {
     display: flex;
     position: fixed;
     top: 50%;
@@ -21,8 +21,8 @@
     border-radius: 8px;
     border: solid 0.5px rgba(255, 255, 255, 0.5);
     transform: translate(-50%, -50%);
-    .compare-header {
-        .compare-title {
+    .Overlap-header {
+        .Overlap-title {
             display: flex;
             align-items: center;
         }
@@ -38,14 +38,38 @@
             margin-right: 4px;
         }
     }
-    .compare-body {
+    .Overlap-body {
         width: 100%;
         height: 90%;
         display: flex;
+        flex-direction: column;
         align-items: center;
         justify-content: center;
     }
-    .compare-footer {
+    .navi-list {
+        font-size: 14px;
+        padding: 8px 0;
+        border-bottom: 1px solid #777777;
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        width: 80%;
+        .navi-button {
+            display: flex;
+            align-items: center;
+            width: calc(30% + 12px);
+            gap: 4px;
+            justify-content: end;
+        }
+        .navi-name {
+            width: 20%;
+        }
+        .navi-time {
+            width: 50%;
+        }
+    }
+
+    .Overlap-footer {
         height: 10%;
         width: 100%;
     }
